backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi
uvicorn
openai-whisper

# CPU-only torch (NO CUDA)
torch==2.2.2+cpu

# CRITICAL: Whisper + Torch require NumPy 1.x
numpy<2

ffmpeg-python
python-multipart
